+++ packages/components/src/local/form-elements/order_detail/index.tsx
@@ -2,7 +2,7 @@ import { faEdit } from '@fortawesome/free-solid-svg-icons'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { Card, CardContent, Grid, Link } from '@material-ui/core'
 import { PlannedActivityGeometry, PlanningActivity, PlanningActivityGeometry, PlatformTypeData } from '@serge/custom-types'
-import { PlanningMessageStructureCore } from '@serge/custom-types/message'
+import { PlanningMessageStructureCore } from '@serge/custom-types/p9legacy'
 import { findAsset } from '@serge/helpers'
 import React, { useEffect, useState } from 'react'
 import styles from './styles.module.scss'
